% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChaosGameTotem3D.R
\name{game3D}
\alias{game3D}
\title{ChaosGameTotem3D}
\usage{
game3D(initial, q, iteration = 20000)
}
\arguments{
\item{initial}{在平面上任意一個初始點及三個(含)以上頂點，為矩陣型態。}

\item{q}{步長比例，設定於[-0.5,0.5]範圍裡，為數值型態。}

\item{iteration}{迭代次數，預設=20000。}
}
\description{
進行Chaos Game，依照package使用者在三維所設置的初始點與隨機的頂點及所走的步長，迭代數次後點出一個圖騰。
}
\details{
Chaos Game是依照玩家給予在三維上一個初始點與任意三個(含)以上的頂點及步長比例。利用隨機選擇玩家給予的任一個頂點，並將其與初始點的距離依照步長比例，畫出一個點，此點即為下一個起點。反覆這些步驟，迭代多次即可畫出明顯的圖騰。
}
\examples{
initial = matrix(c(2,2,2,20,40,0,55,100,0,80,25,0,100,85,0,40,30,20),nrow = 6,ncol = 3,byrow = TRUE)
#初始點(2,2,2)、四個頂點(20,40,0)、(55,100,0)、(80,25,0)、(40,30,20)
game3D(initial,q=0.3)
# returns 一個三角柱圖騰
}
